Added apache license 2.0 files.
To use this license we MUST also provide a NOTICE file that contains brief
information about the project copyright. Also information about any
dependencies we include DIRECTLY in the source, that is, we do not have to
include information about any dependencies the user has to download himself in
order to compile the code. So, remember to update the notice file.